+++ styleguide.config.js
@@ -48,6 +48,7 @@ module.exports = {
       path.join(__dirname, 'src/components/cell-control/input.jsx'),
       path.join(__dirname, 'src/components/cell-control/money.jsx'),
       path.join(__dirname, 'src/components/cell-control/multi-autocompleter.jsx'),
+      path.join(__dirname, 'src/components/cell-control/multi-choice.jsx'),
       path.join(__dirname, 'src/components/cell-control/multi-select.jsx'),
       path.join(__dirname, 'src/components/cell-control/number.jsx'),
       path.join(__dirname, 'src/components/cell-control/select.jsx'),